+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 * Warnings written by `Import-Module` are now hidden.
 * Fixed: `Resolve-WhiskeyTaskPath` can return extra files if searching on a case-sensitive file system and the directory
 being searched has no upper or lower case letters.
+* Fixed: the `DotNet` command fails when running some commands under .NET 6.0 SDK because .NET 6.0 is stricter about
+validating parameters.
+* Added a `NoLog` parameter to the `DotNet` task to turn off logging.
 
 # 0.48.3